+++ src/RootTypeGenerator.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Fluwork\CodeGenerator;
+
+class RootTypeGenerator extends TypeGenerator
+{
+    public function __construct(string $name, private bool $nullable = false)
+    {
+        parent::__construct($name);
+    }
+
+    public function generate(): string
+    {
+        return sprintf('%s%s', $this->nullable ? '?' : '', parent::generate());
+    }
+}