+++ src/TypeGenerator.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Fluwork\CodeGenerator;
+
+class TypeGenerator implements GeneratorInterface
+{
+    public function __construct(private string $type)
+    {
+    }
+
+    public function generate(): string
+    {
+        return $this->type;
+    }
+}