+++ src/Visibility.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Fluwork\CodeGenerator;
+
+class Visibility
+{
+    public const PUBLIC = 'public';
+    public const PROTECTED = 'protected';
+    public const PRIVATE = 'private';
+}