+++ CHANGELOG.md
@@ -8,3 +8,10 @@ This changelog follows [**Keep a changelog** conventions](https://keepachangelog
 
 - `FileGenerator` used to generate PHP files
 - `BlockGenerator` used to generate a block of code
+- `ParameterGenerator` used to generate a function parameter
+- `ParametersGenerator` used to generate a sequence of parameters
+- `RootTypeGenerator` used to generate a type which can be nullable
+- `TypeGenerator` used to generate a plain type
+- `UnionTypeGenerator` used to generate a union type
+- `ValueGenerator` used to generate a string from a mixed value
+- `Visibility` which defines the three possible visibilities in PHP